languages/haskell: hls-wrapper -> hls
hls-wrapper is a simple script to start the correct language server
based on the currently oppened project. From nvf perspective, this makes
it effectively useless. To allow haskell to work on nvf OOTB, we specify
cmd with hls (not wrapper).

NOTE: this pins the language server to specific GHC version.
To circumvent this, users must override (lib.mkForce)
vim.lsp.servers.haskell-tools.cmd with their own, or just specify it as
[ "haskell-language-server-wrapper" "--lsp" ].